# infrastructure/resources/devops-api/docs/swagger.yaml
info:
  contact: {}
paths:
...
  /databases:
    get:
      description: List databases for a service and environment
      parameters:
      - description: Service name
        in: query
        name: service
        required: true
        type: string
      - description: Environment
        in: query
        name: env
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get databases
      tags:
      - databases
...
  /schemas:
    get:
      description: List schemas for a database
      parameters:
      - description: Service name
        in: query
        name: service
        required: true
        type: string
      - description: Environment
        in: query
        name: env
        required: true
        type: string
      - description: Database name
        in: query
        name: db
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get schemas
      tags:
      - schemas
...
swagger: "2.0"